consolidated shared preferences
